requirements.txt

pandas
tk
scikit-learn
pillow
pydot
